refactor: extract battle runner helper
