+++ discord-bot/src/commands/practice.js
@@ -7,9 +7,8 @@ const {
 } = require('discord.js');
 const userService = require('../utils/userService');
 const abilityCardService = require('../utils/abilityCardService');
-const { buildBattleEmbed } = require('../utils/embedBuilder');
+const { runBattleLoop, sendBattleLogDM, formatLog } = require('../utils/battleRunner');
 
-const MAX_LOG_LINES = 20;
 const GameEngine = require('../../../backend/game/engine');
 const { createCombatant } = require('../../../backend/game/utils');
 const { allPossibleHeroes, allPossibleAbilities } = require('../../../backend/game/data');
@@ -25,25 +24,6 @@ function respond(interaction, options) {
 const data = new SlashCommandBuilder()
   .setName('practice')
   .setDescription('Test your skills against a goblin with no risk.');
-
-function formatLog(entry) {
-  const prefix = `[R${entry.round}]`;
-  let text = entry.message;
-
-  switch (entry.type) {
-    case 'round':
-      return `\n**--- ${text} ---**\n`;
-    case 'ability-cast':
-      return `\`\`\`diff\n+ ${prefix} ${text}\n\`\`\``;
-    case 'victory':
-    case 'defeat':
-      return `🏆 **${text}** 🏆`;
-    case 'status':
-      return `\`\`\`css\n. ${prefix} ${text}\n\`\`\``;
-    default:
-      return `${prefix} ${text}`;
-  }
-}
 
 async function execute(interaction) {
   let user = await userService.getUser(interaction.user.id);
@@ -120,23 +100,7 @@ async function execute(interaction) {
   });
 
   const engine = new GameEngine([player, goblin]);
-  const wait = ms => new Promise(r => setTimeout(r, ms));
-  let battleMessage;
-  let logText = '';
-  const fullLog = [];
-  for (const step of engine.runGameSteps()) {
-    fullLog.push(...step.log);
-    const summaryLog = fullLog.filter(entry => entry.level === 'summary');
-    const lines = summaryLog.map(formatLog);
-    logText = lines.slice(-MAX_LOG_LINES).join('\n');
-    const embed = buildBattleEmbed(step.combatants, logText);
-    if (!battleMessage) {
-      battleMessage = await interaction.followUp({ embeds: [embed] });
-    } else {
-      await wait(1000);
-      await battleMessage.edit({ embeds: [embed] });
-    }
-  }
+  const { fullLog } = await runBattleLoop(interaction, engine);
 
   const summaryEmbed = new EmbedBuilder()
     .setColor(engine.winner === 'player' ? '#57F287' : '#ED4245')
@@ -144,53 +108,7 @@ async function execute(interaction) {
 
   await interaction.followUp({ embeds: [summaryEmbed] });
 
-  const finalLogString = fullLog
-    .map(entry => {
-      let prefix = `[R${entry.round}]`;
-      let message = entry.message;
-      switch (entry.type) {
-        case 'round':
-          return `\n--- ${message} ---\n`;
-        case 'ability-cast':
-          message = `✨ ${message}`;
-          break;
-        case 'defeat':
-        case 'victory':
-          message = `🏆 ${message} 🏆`;
-          break;
-        case 'status':
-          message = `💀 ${message}`;
-          break;
-      }
-      return `${prefix} ${message}`.trim();
-    })
-    .join('\n');
-
-  const logBuffer = Buffer.from(finalLogString, 'utf-8');
-
-  if (user.dm_battle_logs_enabled) {
-    try {
-      if (typeof interaction.user.send === 'function') {
-        await interaction.user.send({
-          content: 'Here is the full transcript of your last battle:',
-          files: [{ attachment: logBuffer, name: `battle-log-${Date.now()}.txt` }]
-        });
-      } else {
-        throw new Error('DM function unavailable');
-      }
-    } catch (error) {
-      console.error(`Could not send battle log DM to ${interaction.user.tag}.`, error);
-      await interaction.followUp({
-        content:
-          "I couldn't DM you the full battle log. Please check your privacy settings if you'd like to receive them in the future.",
-        ephemeral: true
-      });
-    }
-  } else {
-    console.log(
-      `[DM DISABLED] Skipping battle log DM for ${interaction.user.username} (dm_battle_logs_enabled = false)`
-    );
-  }
+  await sendBattleLogDM(interaction, user, fullLog);
 }
 
 module.exports = { data, execute };
